requirements.txt
streamlit
google-generativeai
python-dotenv
langchain==0.3.25
PyPDF2
chromadb
faiss-cpu
langchain_google_genai==2.0.1
pdfplumber
easyocr
pdf2image
opencv-python
langchain-core==0.3.75
langchain-community==0.3.24
langchain-text-splitters==0.3.11
sentence-transformers
langchain-huggingface